% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{ad_get_genome}
\alias{ad_get_genome}
\title{Genome information for transcript set}
\usage{
ad_get_genome(details = FALSE)
}
\arguments{
\item{details}{Logical. If TRUE, return seqinfo object. Default: FALSE.}
}
\value{
String (if details == FALSE). Seqinfo object otherwise
}
\description{
Genome information for transcript set
}
