% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{ad_get_packagename}
\alias{ad_get_packagename}
\title{Query name of current data annotation package}
\usage{
ad_get_packagename()
}
\value{
String. Name of annotation data pacakge.
}
\description{
Query name of current data annotation package
}
