% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors.r
\name{ad_is_single_exn_tx}
\alias{ad_is_single_exn_tx}
\title{Query for single exon transcripts}
\usage{
ad_is_single_exn_tx(txname)
}
\arguments{
\item{txname}{String. Name of transcript}
}
\value{
Logical. TRUE if txname is a single exon transcript.
}
\description{
Query for single exon transcripts
}
\details{
Caveat: This function returns "FALSE" for single exon transcripts that are not in the transcript set.
}
