% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swap_annotations.R
\name{ad_swap_annotation}
\alias{ad_swap_annotation}
\title{Swap out annotation package used by aenmd}
\usage{
ad_swap_annotation(PACKAGENAME)
}
\arguments{
\item{PACKAGENAME}{String. Name of the new annotation package to be used.}
}
\value{
None. This function is called only for the side effect of swapping
out the annotation package aenmd is using.
}
\description{
Swap out annotation package used by aenmd
}
