% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmd.R
\name{annotate_nmd_v2}
\alias{annotate_nmd_v2}
\title{Wrapper function for annotating a set of variants with NMD escape}
\usage{
annotate_nmd_v2(vcf_rng, check_ref = FALSE, verbose = FALSE)
}
\arguments{
\item{vcf_rng}{GRanges object of variants. Start is first base of variant.
Each variant is of length 1.
Each variant has an ``ref`` and ``alt`` metadata column with DNAStrings of the sequences.}

\item{check_ref}{Logical. Should variants be verified against refernce sequence.}

\item{verbose}{Logical. Report progress.}
}
\value{
List.
}
\description{
Wrapper function for annotating a set of variants with NMD escape
}
