% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variant.R
\name{annotate_variant}
\alias{annotate_variant}
\title{Workhorse function annotating a variant with escape from nonsense mediated decay}
\usage{
annotate_variant(rng, tx)
}
\arguments{
\item{rng}{GRanges object. Start is first base of variant, end is last base of variant.
Needs an ``ref`` and ``alt`` metadata column.}

\item{tx}{String. The transcript ID for which the variant is annotated.}
}
\value{
List.
}
\description{
Workhorse function annotating a variant with escape from nonsense mediated decay
}
