% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variant.R
\name{annotate_variants_by_tx}
\alias{annotate_variants_by_tx}
\title{Annotate variants overlapping a transcript}
\usage{
annotate_variants_by_tx(txname, vars, detailed = FALSE)
}
\arguments{
\item{txname}{Character. Transcript name.}

\item{vars}{GRanges. Overlapping variants.}

\item{detailed}{Logical. Should additional information be returned.}
}
\value{
GRanges.
}
\description{
Annotate variants overlapping a transcript
}
