% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmd_escape_rules.R
\name{apply_nmd_escape_rules}
\alias{apply_nmd_escape_rules}
\title{Determine NMD escape rules}
\usage{
apply_nmd_escape_rules(
  ptc_loc,
  exn_ind,
  exn_sta,
  exn_end,
  num_exn,
  txname,
  css_prox_dist = 150L,
  penultimate_prox_dist = 50L
)
}
\arguments{
\item{ptc_loc}{Integer. Location of PTC. In CDS codon coordinates (i.e., amino acids), 1-based.}

\item{exn_ind}{Integer. Index of the exon the variant overlaps, 1-based.}

\item{exn_sta}{Integer. Start of the exon. In TX coordinates (i.e., nucleotides), 1-based, inclusive (i.e., first nucleotide of the exon).}

\item{exn_end}{Integer. End of the exon. In TX coordinates, 1-based, inclusive (i.e., last nucleotide of the exon).}

\item{num_exn}{Integer. Number of exons in the CDS of the transcript (i.e., coding exons).}

\item{txname}{Character. Ensembl transcript id.}

\item{css_prox_dist}{Integer. Distance cutoff (strict, <) for NMD-escape region near the coding start site. Default: 150.}

\item{penultimate_prox_dist}{Integer. Distance cutoff (strict, <) for NMD-escape region neat the 3'-boundary of the penultimate exon. Default: 50.}
}
\value{
Logical. Vogical vector with six named elements.
}
\description{
Determine NMD escape rules
}
\details{
Return value (Logical vector) has six named elements.
- \code{is_ptc = TRUE} if there is a PTC in the alternative transcript. 
- \code{is_last = TRUE} if the PTC is in the last (3'-most) coding exon.
- \code{is_penultimate3PP = TRUE} if the PTC is within the last (i.e., downstream, 3'-most) \code{penultimate_prox_dist}bp of the penultimate exon.
- \code{is_css5PP  = TRUE} if the PTC is within the first (upstream, 5'-most) \code{css_prox_dist} of the translation/coding start site. 
- \code{is_single = TRUE} if the PTC is in a single exon transcript.
- \code{is_407plus = TRUE} if the PTC is in an exon that is longer than 407bp.
}
