% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dist_downstream}
\alias{get_dist_downstream}
\title{Get distance between a PTC (or any AA) and a downstream boundary (e.g., exon end)}
\usage{
get_dist_downstream(aa_pos, ds_bnd)
}
\arguments{
\item{aa_pos}{Integer. Location of the AA/PTC in the CDS. CDS coordinates (codonn), 1-based.}

\item{ds_bnd}{Integer. Location of the last nucleotide before the boundary (e.g., last nucleotide in exon), 1-based.}
}
\value{
Integer. Distance of the AA/PTC (acutally, the downstream-most nucleotide covered by the AA/PTC to the boundary).
}
\description{
Get distance between a PTC (or any AA) and a downstream boundary (e.g., exon end)
}
\details{
\preformatted{
NUC:  123456789...  ( '|' = exon-exon boundary (nucleotide to the left is last of upstream exon), X = PTC-covered, O = non-PTC-covered)
AA:   111222333
PTC:  ATGOOOOOOX|XXOOOOOOOOOOOO  aa_pos: 4, ds_bnd: 10, DIST: 0
PTC:  ATGOOOOOOXX|XOOOOOOOOOOOO  aa_pos: 4, ds_bnd: 11, DIST: 0
PTC:  ATGOOOOOOXXX|OOOOOOOOOOOO  aa_pos: 4, ds_bnd: 12, DIST: 0
PTC:  ATGOOOOOOXXXO|OOOOOOOOOOO  aa_pos: 4, ds_bnd: 13, DIST: 1
PTC:  ATGOOOOOOXXXOO|OOOOOOOOOO  aa_pos: 4, ds_bnd: 14, DIST: 2
PTC:  ATGOOOOOOXXXOOO|OOOOOOOOO  aa_pos: 4, ds_bnd: 15, DIST: 3
}
}
