% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dist_upstream}
\alias{get_dist_upstream}
\title{Get distance between a PTC (or any AA) and a upstream boundary (e.g., a CSS site)}
\usage{
get_dist_upstream(aa_pos, us_bnd)
}
\arguments{
\item{aa_pos}{Integer. Location of the AA/PTC in the CDS. CDS coordinates (codon), 1-based.}

\item{us_bnd}{Integer. Location of the last nucleotide before the boundary (e.g., first nucleotide in exon), 1-based.}
}
\value{
Integer. Distance of the AA/PTC (acutally, the upstream-most nucleotide covered by the AA/PTC to the boundary).
}
\description{
Get distance between a PTC (or any AA) and a upstream boundary (e.g., a CSS site)
}
\details{
\preformatted{
NUC:  123456789...  ( '|' = exon-exon boundary (nucleotide to the left is last of upstream exon), X = PTC-covered, O = non-PTC-covered)
PTC  |ATGXXXOOOOOOOOO  aa_pos: 2, us_bnd: 1, DIST: 3
PTC  |ATGOOOXXXOOOOOO  aa_pos: 3, us_bnd: 1, DIST: 6
PTC  |ATGOOOOOOXXXOOO  aa_pos: 4, us_bnd: 1, DIST: 9
PTC  |ATGOOOOOOOOOXXX  aa_pos: 5, us_bnd: 1, DIST: 12

NUC:  123456789 
PTC:  ATGOOOOOOXX|XOOOOOOOOOOO  aa_pos: 4, us_bnd: 12, DIST: 0
PTC:  ATGOOOOOOX|XXOOOOOOOOOOO  aa_pos: 4, us_bnd: 11, DIST: 0
PTC:  ATGOOOOOO|XXXOOOOOOXXXOO  aa_pos: 4, us_bnd: 10, DIST: 0
PTC:  ATGOOOOO|OXXXOOOOOOXXXOO  aa_pos: 4, us_bnd: 9,  DIST: 1
PTC:  ATGOOOO|OOXXXOOOOOOXXXOO  aa_pos: 4, us_bnd: 8,  DIST: 2
PTC:  ATGOOO|OOOXXXOOOOOOXXXOO  aa_pos: 4, us_bnd: 7,  DIST: 3
}
}
