% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variant.R
\name{get_rules}
\alias{get_rules}
\title{Determine NMD escape rules}
\usage{
get_rules(ptc_loc, exn_ind, exn_sta, exn_end, num_exn)
}
\arguments{
\item{ptc_loc}{Integer. Location of PTC. In CDS codon coordinates, 1-based.}

\item{exn_ind}{Integer. Index of the exon the variant overlaps, 1-based.}

\item{exn_sta}{Integer. Start of the exon. In CDS codon coordinates, 1-based.}

\item{exn_end}{Integer. End of the exon. In CDS codon coordinates, 1-based.}

\item{num_exn}{Integer. Number of exons in transcript.}
}
\value{
Logical. Vogical vector with six named elements.
}
\description{
Determine NMD escape rules
}
\details{
Return value (Logical vector) has six named elements.
- \code{is_ptc = TRUE} if there is a PTC in the alternative transcript. 
- \code{is_last = TRUE} if the PTC is in the last (3'-most) coding exon.
- \code{is_penultimate = TRUE} if the PTC is in the last 50bp of the penultimate exon.
- \code{is_first = TRUE} if the PTC is in the first (5'-most) 150bp of the first (coding) exon.
- \code{is_single = TRUE} if the PTC is in a single exon transcript.
- \code{is_407plus = TRUE} if the PTC is in an exon that is longer than 407bp.
}
