% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_start_end}
\alias{get_start_end}
\title{Variant coordinates in CDS space}
\usage{
get_start_end(gr1, gr2)
}
\arguments{
\item{gr1}{GRanges object. Single range, no strand. The variant. Start is first nucleotide, end is last nucleotide.}

\item{gr2}{GRanges object. Exon ranges. All on the same strand. Sorted 5' to 3' (i.e., "reversed" on negative strand).
Both ``gr1`` and ``gr2`` need to be on the same chromosome.}
}
\value{
Numeric. Vector with start, end, and the overlapping exons.
}
\description{
Get first and last base of a variant in a transcript's CDS coordinates
}
\details{
|123|---|4567|-----|89|  <- positive strand
     |xxx|---|xxxx|-----|xx|
       ===================    gr1 => (2,8)
       **     ****       *    intersect(gr1,gr2)

      |987|---|6543|-----|21|  <- negative strand
      |xxx|---|xxxx|-----|xx|
         ===============       gr1 ==> (3,7)
         *     ****
}
\examples{
gr1  <- GenomicRanges::GRanges('chr1:2-18:*')
gr21 <- GenomicRanges::GRanges('chr1:1-3:+')
gr22 <- GenomicRanges::GRanges('chr1:7-10:+')
gr23 <- GenomicRanges::GRanges('chr1:16-20:+')
gr2  <- c(gr21,gr22,gr23)
aenmd:::get_start_end(gr1,gr2) #- 2, 10, 1, 3
}
