% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mev_alt_exn_bnd}
\alias{mev_alt_exn_bnd}
\title{Update exon boundaries for alternative allele's sequence}
\usage{
mev_alt_exn_bnd(exn_sta_t, exn_end_t, exn_inds, d_w, sci)
}
\arguments{
\item{exn_sta_t}{Integer. Vector, the exon starts (in transcript coordinates), reference allele}

\item{exn_end_t}{Integer. Same as \code{exn_sta_t}, just the exon ends}

\item{exn_inds}{Integer. The exons (typically one, can be more than one) the variant overlaps}

\item{d_w}{Integer. The diffrence in length between reference and alterntive allele.}

\item{sci}{Integer. The start in the alternative sequence where its start codon starts.}
}
\value{
List. Two components, \code{exn_sta_t_alt} and \code{exn_end_t_alt}.
}
\description{
Update exon boundaries for alternative allele's sequence
}
