% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_variant.R
\name{normalize_var}
\alias{normalize_var}
\title{Left-normalize genetic variants}
\usage{
normalize_var(rng, ref = NULL, alt = NULL, prefix = NULL)
}
\arguments{
\item{rng}{GRanges object representing a variant. Range here is *first* and *last* base of reference variant.}

\item{ref}{DNAString. The reference allele of the variant}

\item{alt}{DNAString. The alternative allele of the variant}

\item{prefix}{DNAString. Nucleotides preceeding the variant, used for normalization.}
}
\value{
The normalized variant (as GRanges object)
}
\description{
Convert variants into left-normalized version (PMID: 25701572)
}
\examples{
prefix     <- Biostrings::DNAString("GATAT")
ref        <- Biostrings::DNAString("GAGATAT")
alt        <- Biostrings::DNAString("GAGAT")
start      <- 17
end        <- 17+6
rng        <- GenomicRanges::GRanges("1",IRanges::IRanges(17,17+6))
rng$ref    <- Biostrings::DNAStringSet(ref)
rng$alt    <- Biostrings::DNAStringSet(alt)
rng$prefix <- Biostrings::DNAStringSet(prefix)
normalize_var(rng, rng$ref, rng$alt, rng$prefix)
}
