% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_variant.R
\name{normalize_vars}
\alias{normalize_vars}
\title{Left-normalize genetic variants}
\usage{
normalize_vars(rng)
}
\arguments{
\item{rng}{GRanges object representing a variant. Range here is *first* and *last* base of refernce variant.
Needs to have ``ref``, ``alt`` and ``prefix`` in the metadata.}
}
\value{
The normalized variant (as GRanges object)
}
\description{
Convert variants into left-normalized version (PMID: 25701572);
Naive (i.e., ,slow) "vectorization" of the ``normalize_var`` function.
}
