% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmd.R
\name{parse_vcf}
\alias{parse_vcf}
\title{Read in vcf file, make GRanges object.}
\usage{
parse_vcf(vcf_filename, pass_only = TRUE)
}
\arguments{
\item{vcf_filename}{String. Filename of vcf file.}

\item{pass_only}{Logical. Keep only entries where FILTER = PASS}
}
\description{
Uses vcfR package. Keep only bare-bone info from vcf.
}
