% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmd.R
\name{annotate_nmd}
\alias{annotate_nmd}
\title{Wrapper function for annotating a set of variants with NMD escape}
\usage{
annotate_nmd(vcf_rng, check_ref = TRUE)
}
\arguments{
\item{vcf_rng}{GRanges object of variants. Start is first base of variant.
Each variant is of length 1.
Each variant has an ``ref`` and ``alt`` metadata column with DNAStrings of the sequences.}

\item{check_ref}{Logical. Should the reference alleles be checked against assembly.}
}
\value{
List.
}
\description{
Wrapper function for annotating a set of variants with NMD escape
}
