% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_nmd.R
\name{annotate_nmd_v2}
\alias{annotate_nmd_v2}
\title{Wrapper function for annotating a set of variants with NMD escape}
\usage{
annotate_nmd_v2(
  vcf_rng,
  check_ref = FALSE,
  verbose = FALSE,
  multicore = FALSE,
  num_cores = 2,
  rettype = "grl"
)
}
\arguments{
\item{vcf_rng}{GRanges object of variants. Start is first base of variant.
Each variant is of length 1.
Each variant has an ``ref`` and ``alt`` metadata column with DNAStrings of the sequences.}

\item{check_ref}{Logical. Should variants be verified against refernce sequence.}

\item{verbose}{Logical. Report progress.}

\item{multicore}{Logical. Should multiple cores be used in the computations (via the parallel pacakge).}

\item{num_cores}{Integer. The number of cores to use. Only when \code{multicore = TRUE}.}

\item{rettype}{Character. Should a `GRangesList` be returned (default, `rettype = 'grl'`) or shold results be collated into a `GRanges` object (`rettype = 'gr'`).}
}
\value{
List.
}
\description{
Wrapper function for annotating a set of variants with NMD escape
}
\details{
For multicore, the `BiocParallel` default backand returned by `BiocParallel::bpparam()` is used.
}
