% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_variant.R
\name{get_rules}
\alias{get_rules}
\title{Determine NMD escape rules}
\usage{
get_rules(ptc_loc, exn_ind, exn_sta, exn_end, num_exn)
}
\arguments{
\item{ptc_loc}{Integer. Location of PTC. In CDS codon coordinates, 1-based.}

\item{exn_ind}{Integer. Index of the exon the variant overlaps, 1-based.}

\item{exn_sta}{Integer. Start of the exon. In CDS codon coordinates, 1-based.}

\item{exn_end}{Integer. End of the exon. In CDS codon coordinates, 1-based.}

\item{num_exn}{Integer. Number of exons in transcript.}
}
\description{
Determine NMD escape rules
}
