% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_io.R
\name{parse_vcf_VariantAnnotation}
\alias{parse_vcf_VariantAnnotation}
\title{Simple wrapper to read in in vcf file, make GRanges object.}
\usage{
parse_vcf_VariantAnnotation(vcf_filename, verbose = TRUE, ...)
}
\arguments{
\item{vcf_filename}{String. Filename of vcf file.}

\item{verbose}{Logical. Report progress.}

\item{...}{dotdotdot. Passed to \code{VariantAnnotation::readVcf}.}
}
\value{
List. \code{vcf_rng} contains the fixed part of the vcf-file as GRanges. \code{vcf} is the complete \code{vcfR::vcfR} object.
}
\description{
Uses VariantAnnotation package.
}
